separte test code, connect db, create preprints
