+++ kisti/osf/data_object.py
@@ -1,0 +1,13 @@
+import json
+
+
+def get_json_id(data):
+    data = json.loads(data)
+    id = data.get("data").get("id")
+    return id
+
+
+def get_files_id(data):
+    data = json.loads(data)
+    id = data.get('data').get('id')
+    return id.lstrip("osfstorage/")
